Creates function to give winner of total games
Previous code was only suited for one round. So this function just
plays the round of 5 games and give the overall winner of the game.

Also previously the function were called right after the function block
but doing that caused problem with the counter of scores so, now
all the function are called properly inside a playGame function making
them the one part of the game.